requirements.txt
azure-functions==1.11.0
requests==2.26.0
azure-storage-blob==12.9.0